- Fixed some more issues where cancels were incorrectly observed as "set checksum to None" by the high level - Added a small delay for auth-related structured cell joins, to give them some time so that they might be updated without canceling by the next join.

+++ seamless/core/manager/cancel.py
@@ -630,8 +630,9 @@ class CancellationCycle:
 
         for macro in macros_to_destroy:
             gen_context = macro._gen_context
-            gen_context.destroy()
-            macro._gen_context = None
+            if gen_context is not None:
+                gen_context.destroy()
+                macro._gen_context = None
 
 
 from ..utils import overlap_path
